# docs/swagger.yaml
basePath: /
definitions:
  models.Segment:
    properties:
      slug:
        type: string
    type: object
  segment.CreateRequest:
    properties:
      slug:
        type: string
    type: object
  segment.DeleteRequest:
    properties:
      slug:
        type: string
    type: object
  segment.GetSegmentsForUserRequest:
    properties:
      user_id:
        type: integer
    type: object
  segment.HistoryRequest:
    properties:
      month:
        type: integer
      user_id:
        type: integer
      year:
        type: integer
    type: object
  segment.UpdateUserSegmentsRequest:
    properties:
      add_segments:
        items:
          type: string
        type: array
      delete_segments:
        items:
          type: string
        type: array
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Тествое задание для Avitotech 2023
  title: Avitotech Test 2023 API
  version: "1.0"
paths:
  /segment:
    delete:
      consumes:
      - application/json
      description: Метод удаления сегмента. Принимает slug (название) сегмента.
      parameters:
      - description: Segment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/segment.DeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete segment
      tags:
      - segment
    post:
      consumes:
      - application/json
      description: Метод создания сегмента. Принимает slug (название) сегмента.
      parameters:
      - description: Segment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/segment.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              created_at:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create segment
      tags:
      - segment
  /segment/reports/{fileName}:
    get:
      description: Метод скачивания csv отчета по истории сегментов пользователя.
      parameters:
      - description: название файла в формате /reports/file_name.csv
        in: path
        name: fileName
        required: true
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          headers:
            Content-Disposition:
              description: attachment;filename=file_name
              type: string
            Content-Type:
              description: text/csv
              type: string
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Скачивание отчета
      tags:
      - segment
  /segment/user:
    get:
      consumes:
      - application/json
      description: Метод получения активных сегментов пользователя. Принимает на вход
        id пользователя.
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/segment.GetSegmentsForUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Segment'
            type: array
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get user segments
      tags:
      - segment
    post:
      consumes:
      - application/json
      description: Метод добавления пользователя в сегмент. Принимает список slug
        (названий) сегментов которые нужно добавить пользователю,
      parameters:
      - description: Segment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/segment.UpdateUserSegmentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              segments_added:
                type: integer
              segments_deleted:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update user segments
      tags:
      - segment
  /segment/user/history:
    get:
      consumes:
      - application/json
      description: 'Метод получения истории сегментов пользователя за указанный месяц
        и год. На вход: год и месяц. На выходе ссылка на CSV файл.'
      parameters:
      - description: Segment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/segment.HistoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              report:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get user segments history
      tags:
      - segment
swagger: "2.0"